[STMT-102] 🐛 로그인과 회원가입 전환 시 발생하는 앱 강제 종료 현상 수정
로그인과 회원가입이 동일한 네비게이션 스택 내에서 이동하는 구조이기에, 로그인과 회원가입 코디네이터를 하나의 통합 코디네이터로 변경하였습니다.
