+++ Stumeet/Stumeet/Presentation/AuthScenes/Register/ViewModel/NicknameViewModel.swift
